# test(client): Optimize integration tests