--- a/packages/client/test/integration/GapFill.test.ts
+++ b/packages/client/test/integration/GapFill.test.ts
@@ -5,7 +5,6 @@ import { StreamrClientConfig } from '../../src/Config'
 import { StreamPermission } from '../../src/permission'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
-import { Subscriber } from '../../src/subscribe/Subscriber'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import { FakeStorageNode } from '../test-utils/fake/FakeStorageNode'
 import { getPublishTestStreamMessages, Msg } from '../test-utils/publish'
@@ -38,7 +37,6 @@ describe('GapFill', () => {
     let publishTestMessages: ReturnType<typeof getPublishTestStreamMessages>
     let client: StreamrClient
     let stream: Stream
-    let subscriber: Subscriber
     let storageNode: FakeStorageNode
     let environment: FakeEnvironment
 
@@ -49,8 +47,6 @@ describe('GapFill', () => {
             retryResendAfter: 1000,
             ...opts
         })
-        // @ts-expect-error private
-        subscriber = client.subscriber
         stream = await createTestStream(client, module)
         await stream.grantPermissions({ permissions: [StreamPermission.SUBSCRIBE], public: true })
         await stream.addToStorageNode(storageNode.id)
@@ -63,7 +59,7 @@ describe('GapFill', () => {
     })
 
     afterEach(async () => {
-        const subscriptions = await subscriber.getSubscriptions()
+        const subscriptions = await client.getSubscriptions()
         expect(subscriptions).toHaveLength(0)
     })
 
@@ -87,7 +83,7 @@ describe('GapFill', () => {
                     return undefined
                 })
 
-                expect(await subscriber.count(stream.id)).toBe(1)
+                expect(await client.getSubscriptions(stream.id)).toHaveLength(1)
 
                 const published = await publishTestMessages(MAX_MESSAGES)
 
@@ -112,7 +108,7 @@ describe('GapFill', () => {
                     return undefined
                 })
 
-                expect(await subscriber.count(stream.id)).toBe(1)
+                expect(await client.getSubscriptions(stream.id)).toHaveLength(1)
 
                 const published = await publishTestMessages(MAX_MESSAGES)
 
@@ -134,7 +130,7 @@ describe('GapFill', () => {
                     return undefined
                 })
 
-                expect(await subscriber.count(stream.id)).toBe(1)
+                expect(await client.getSubscriptions(stream.id)).toHaveLength(1)
 
                 const published = await publishTestMessages(MAX_MESSAGES)