--- a/packages/client/test/integration/Resends2.test.ts
+++ b/packages/client/test/integration/Resends2.test.ts
@@ -42,14 +42,14 @@ describe('Resends2', () => {
     })
 
     beforeEach(async () => {
-        client = environment.createClient()
-        stream = await createTestStream(client, module)
-        await stream.grantPermissions({
-            user: publisherWallet.address,
-            permissions: [StreamPermission.PUBLISH]
+        stream = await createTestStream(publisher, module)
+        await publisher.grantPermissions(stream.id, {
+            public: true,
+            permissions: [StreamPermission.SUBSCRIBE]
         })
         storageNode = environment.startStorageNode()
         await stream.addToStorageNode(storageNode.id)
+        client = environment.createClient()
     })
 
     afterEach(async () => {
@@ -534,7 +534,7 @@ describe('Resends2', () => {
         const publishedMessage = Msg({
             content: fs.readFileSync(path.join(__dirname, '../data/utf8Example.txt'), 'utf8')
         })
-        const publishReq = await client.publish(stream, publishedMessage)
+        const publishReq = await publisher.publish(stream, publishedMessage)
 
         await getWaitForStorage(client)(publishReq)
         const sub = await client.resend(stream.id,